requirements.txt
pyobjc-core==3.2.1
pyobjc-framework-Quartz==3.2.1
